# docs/swagger.yml
openapi: 3.0.0
info:
  title: Merchant API
  version: '1.0'
  description: |-
    Nuestra API de gestión de comercios permite a los desarrolladores integrar fácilmente la gestión de inventario, pedidos y clientes en sus aplicaciones. 
    
    Con nuestra API, los comerciantes pueden dar de alta sus comercios y usuarios, y los usuarios pueden visitar los comercios y dejar sus valoraciones. 
    Además, los comercios pueden ver información sobre los clientes que visiten sus páginas. 
    
    _Esto permite a los comerciantes automatizar sus procesos y mejorar la eficiencia de su negocio._ 
  contact:
    email: alejandro.gallego@live.u-tad.com
  license:
    name: U-tad
    url: https://u-tad.com/
servers:
    - url: http://localhost:3000/api/
...
components:
  securitySchemes:
    bearerAuth:
        type: http
        scheme: bearer
        bearerFormat: JWT  
  schemas:
    Paginas:
      type: object
      properties:
        email_propietario:
          type: string
        titulo:
          type: string
          uniqueItems: true
        ciudad:
          type: string
        actividad:
          type: string
        resumen:
          type: string
        textos:
          type: array
          items:
            type: string
        fotos:
          type: array
          items:
            type: object
            properties:
              filename:
                type: string
              url:
                type: string
        rating:
          type: integer
          default: 0
        reseñas:
          type: array
          items:
            type: object
            properties:
              usuario:
                type: string
              puntuacion:
                type: integer
              comentario:
                type: string
    Reseñia:
                type: object
                properties:
                  puntuacion:
                    type: integer
                  comentario:
                    type: string
...
paths:
...
  /paginas:
    get:
      summary: Muestra todas las paginas 
      tags:
        - "Paginas"
      parameters:
        - name: orden
          in: query
          description: Ordenar páginas por rating
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Paginas'
        '4XX': 
          description: 'ERROR_OBTENER_PAGINAS' 
    post:
      summary: Crear una nueva pagina
      security:
        - bearerAuth: [] 
      requestBody:
        description: La pagina que se va a crear
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Paginas'
      tags:
        - "Paginas"
      responses:
        '200':
          description: "La pagina {titulo} ha sido creada con exito y asignada a {email}"
        '4XX': 
          description: 'ERROR_CREAR_PAGINAS'   
        '401': 
          description: 'ERROR_ID_TOKEN' 
        '402': 
          description: 'NOT_SESSION' 
        '403': 
          description: 'NOT_ALLOWED' 
        '405': 
          description: 'ERROR_PERMISSION' 
    put:
      summary: Actualizar la pagina con el token dado
      security:
        - bearerAuth: [] 
      requestBody:
        description: La pagina que se va a actualizar
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Paginas'
      tags:
        - "Paginas"
      responses:
        '200':
          description: "Página {titulo} actualizada correctamente."
        '4XX': 
          description: 'ERROR_ACTUALIZAR_PAGINAS' 
        '401': 
          description: 'ERROR_ID_TOKEN' 
        '402': 
          description: 'NOT_SESSION' 
        '403': 
          description: 'NOT_ALLOWED' 
        '405': 
          description: 'ERROR_PERMISSION' 
    delete:
      summary: Eliminar la pagina con el token dado
      security:
        - bearerAuth: []
      tags:
        - "Paginas"
      responses:
        '200':
          description: "Página {titulo} eliminada correctamente."
        '4XX': 
          description: 'ERROR_ELIMINAR_PAGINAS' 
        '401': 
          description: 'ERROR_ID_TOKEN' 
        '402': 
          description: 'NOT_SESSION' 
        '403': 
          description: 'NOT_ALLOWED' 
        '405': 
          description: 'ERROR_PERMISSION' 
  /paginas/{titulo}:
    get:
      summary: Muestra la pagina que coincida con el titulo 
      tags:
        - "Paginas"
      parameters:
        - name: titulo
          in: path
          description: El titulo de la pagina a encontrar
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Paginas'
        '4XX': 
          description: 'ERROR_OBTENER_PAGINAS' 
    patch:
      summary: Añadir o modificar una reseña en la web del titulo con el token dado
      security:
        - bearerAuth: []
      parameters:
        - name: titulo
          in: path
          description: El titulo de la pagina en la que se publica la reseña
          required: true
          schema:
            type: string
      requestBody:
        description: La reseña a publicar
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/Reseñia'
      tags:
        - "Paginas"
      responses:
        '200':
          description: "El usuario {email} ha publicado su reseña en {titulo} con exito"
        '4XX': 
          description: 'ERROR_ACTUALIZAR_PAGINAS' 
        '401': 
          description: 'ERROR_ID_TOKEN' 
        '402': 
          description: 'NOT_SESSION' 
        '403': 
          description: 'NOT_ALLOWED' 
        '405': 
          description: 'ERROR_PERMISSION'
...